Ex FitNN: Addressed comments from PR,moved from ex4 folder to ex3
