+++ examples/README.md
@@ -14,9 +14,4 @@
 
 ## Example 3
 
-`ex3/`: use conjgrad for training a NN.
-
-
-## Example 4
-
-`ex4/`: use conjgrad for fitting a NN to target function.
+`ex3/`: use conjgrad for fitting a NN to target function.
